streamlit/pyproject.toml
[project]
name = "jogo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "plotly>=6.3.0",
    "polars>=1.32.3",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "streamlit>=1.48.1",
]
